+++ Render_LL_MMCG/src/Sphere.java
@@ -4,16 +4,16 @@
 //della sfera con un raggio, e la funzione che restituisce
 //la normale alla sfera in un punto dato
 public class Sphere {
-	// raggio
-	public float rad;
-	// posizione (centro)
-	public Point3D p;
+  // raggio
+  public float rad;
+  // posizione (centro)
+  public Point3D p;
 
-	//costruttore
-	public Sphere(float nrad, Point3D np) {
-		rad=nrad;
-		p=np;
-	}
+  //costruttore
+  public Sphere(float nrad, Point3D np) {
+    rad=nrad;
+    p=np;
+  }
 
   //metodo che imposta, a seconda della scelta
   //effettuata dall'utente, la posizione
@@ -47,51 +47,50 @@ public class Sphere {
   }
 
   //funzione di intersezione con un raggio: Return
-	//distanza o -1.0f se non c'e' intersezione
-	double intersect(Ray r){
-		//sostituendo il raggio o+td all'equazione della 
-		//sfera (td+(o-p)).(td+(o-p)) -R^2 =0 si ottiene 
-		//l'intersezione. Si deve risolvere
-		//t^2*d.d + 2*t*(o-p).d + (o-p).(o-p)-R^2 = 0
-		// A=d.d=1 (r.d e' normalizzato)
-		// B=2*(o-p).d
-		// C=(o-p).(o-p)-R^2
-		Point3D op = p.subtract(r.o);
-		// calcolo della distanza
-		double t;
-		// 2*t*B -> semplificato usando b/2 invece di B
-		double B=op.dotProduct(r.d);
-		double C=op.dotProduct(op)-rad*rad;
-		//determinante equazione quadratica
-		double det=B*B-C;
-		// se e' negativo non c'e' intersezione reale,
-		//altrimenti assegna a det il risultato
-		if (det<0) 
-			return -1.0f; 
-		else { 
-			det=(float) Math.sqrt(det);
-	        //ritorna la t piu' piccola se questa e' >0 
-			//altrimenti vedi quella piu' grande se e' >0 
-			//se sono tutte negative non c'e' intersezione
-			if((t=B-det)> Utilities.EPSILON)
-				return t;
-			else {
-				if((t=B+det)> Utilities.EPSILON)
-					return t;
-				else
-					return -1.0f;
-			}
-		}
-	}
+  //distanza o -1.0f se non c'e' intersezione
+  double intersect(Ray r) {
+    //sostituendo il raggio o+td all'equazione della
+    //sfera (td+(o-p)).(td+(o-p)) -R^2 =0 si ottiene
+    //l'intersezione. Si deve risolvere
+    //t^2*d.d + 2*t*(o-p).d + (o-p).(o-p)-R^2 = 0
+    // A=d.d=1 (r.d e' normalizzato)
+    // B=2*(o-p).d
+    // C=(o-p).(o-p)-R^2
+    Point3D op = p.subtract(r.o);
+    // calcolo della distanza
+    double t;
+    // 2*t*B -> semplificato usando b/2 invece di B
+    double B=op.dotProduct(r.d);
+    double C=op.dotProduct(op)-rad*rad;
+    //determinante equazione quadratica
+    double det=B*B-C;
+    // se e' negativo non c'e' intersezione reale,
+    //altrimenti assegna a det il risultato
+    if (det<0)
+      return -1.0f;
+    else {
+      det = Math.sqrt(det);
+      //ritorna la t piu' piccola se questa e' >0
+      //altrimenti vedi quella piu' grande se e' >0
+      //se sono tutte negative non c'e' intersezione
+      if((t=B-det) > Utilities.EPSILON)
+        return t;
+      else {
+        if((t=B+det) > Utilities.EPSILON)
+          return t;
+        else
+          return -1.0f;
+      }
+    }
+  }
 
-	//funzione che calcola la normale in un punto iP della 
-	//sfera
+  //funzione che calcola la normale in un punto iP della sfera
   Point3D normal(Point3D iP) {
-	  //vettore dal centro all'intersezione normalizzato
+    //vettore dal centro all'intersezione normalizzato
     return (iP.subtract(p)).getNormalizedPoint();
-	}
+  }
 
-	void translate(Point3D direction) {
-	  p.add(direction);
+  void translate(Point3D direction) {
+    p.add(direction);
   }
 }
